Migrate class progressions to JSON, add cantrip selection, fix float speeds
Major refactoring moving PHB progressions from hardcoded Python constants to JSON data:

- Migrated ASI levels, spell learning deltas, metamagic counts, invocation progressions, expertise levels, and spell slot tables from code to class JSON files and spell_slot_progressions.json
- Updated GameDataLoader with new accessors (get_asi_levels, get_spell_learn_delta, get_metamagic_count_by_level, etc.)
- Fixed speed/movement_used to use REAL/float type to support fractional movement (e.g., 7.5m for Halfling, 1.5m steps)
- Added cantrip and 1st-level spell selection at character creation for know-classes (Bardo, Stregone, Warlock)
- Added runtime display of Bardic Inspiration die and Sneak Attack dice based on character level
- Added show_error_dialog() UI helper for consistent error feedback
- Fixed control list modifications to use in-place operations instead of reassignment
- Removed duplicate __pycache__ directories and fixed various minor bugs

No functional values changed—all progressions verified against PHB during audit migrations.

+++ data/database.py
@@ -172,7 +172,7 @@ def _create_tables(conn: sqlite3.Connection) -> None:
 
             -- Combattimento
             ac                   INTEGER NOT NULL DEFAULT 10,
-            speed                INTEGER NOT NULL DEFAULT 9,
+            speed                REAL NOT NULL DEFAULT 9,
             hit_dice_type        INTEGER NOT NULL DEFAULT 6,
             hit_dice_total       INTEGER NOT NULL DEFAULT 1,
             hit_dice_remaining   INTEGER NOT NULL DEFAULT 1,
@@ -185,7 +185,7 @@ def _create_tables(conn: sqlite3.Connection) -> None:
             action_used          INTEGER NOT NULL DEFAULT 0,
             bonus_action_used    INTEGER NOT NULL DEFAULT 0,
             reaction_used        INTEGER NOT NULL DEFAULT 0,
-            movement_used        INTEGER NOT NULL DEFAULT 0,
+            movement_used        REAL NOT NULL DEFAULT 0,
             previous_turn_state  TEXT NOT NULL DEFAULT '',
 
             -- Magia
